Split Tailwind sandbox behavior checks
