+++ app/src/styles/global.css
@@ -11,6 +11,8 @@
 
 @source not "./styles.json";
 @source not "../sandbox/ariakit-tailwind";
+@source not "../sandbox/ariakit-tailwind-frame-cover-rtl";
+@source not "../sandbox/ariakit-tailwind-layer-band-boundaries";
 
 @theme {
   --color-canvas: oklch(99.33% 0.0011 197.14);
